ad_detection/requirements.txt
# AD Detection with LLM + Acoustic Features
# Based on: "Integrating Fine-Tuned LLM with Acoustic Features for Enhanced Detection of AD"

# Core ML/DL
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0  # For LoRA fine-tuning
bitsandbytes>=0.41.0  # For quantization
accelerate>=0.25.0
datasets>=2.14.0

# Audio processing
torchaudio>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0

# Speech-to-text
openai-whisper>=20231117
# or: faster-whisper>=0.9.0 (for faster inference)

# Classical ML
scikit-learn>=1.3.0
xgboost>=2.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Optimization
optuna>=3.4.0

# Progress & logging
tqdm>=4.65.0
tensorboard>=2.14.0
wandb>=0.16.0  # optional

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0

